ekyc_ai/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-elasticsearch>=0.2.0
pypdf>=3.17.0
pdf2image>=1.16.0
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python>=4.8.0
elasticsearch>=8.0.0
python-multipart>=0.0.6
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
easyocr>=1.7.0
